Update to storage SDK 11.x - Pull through Host.Storage 4.x, Extensions.Storage 4.x, WebJobs.Extensions 4.x - Add test cases

+++ test/WebJobs.Script.Tests.Integration/Management/FunctionsSyncManagerTests.cs
@@ -290,7 +290,10 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Managment
             using (var env = new TestScopedEnvironmentVariable(_vars))
             {
                 var hashBlob = await _functionsSyncManager.GetHashBlobAsync();
-                await hashBlob.DeleteIfExistsAsync();
+                if (hashBlob != null)
+                {
+                    await hashBlob.DeleteIfExistsAsync();
+                }
 
                 var syncResult = await _functionsSyncManager.TrySyncTriggersAsync(checkHash: true);
                 Assert.True(syncResult.Success);
@@ -345,7 +348,10 @@ namespace Microsoft.Azure.WebJobs.Script.Tests.Managment
             using (var env = new TestScopedEnvironmentVariable(_vars))
             {
                 var hashBlob = await _functionsSyncManager.GetHashBlobAsync();
-                await hashBlob.DeleteIfExistsAsync();
+                if (hashBlob != null)
+                {
+                    await hashBlob.DeleteIfExistsAsync();
+                }
 
                 _mockHttpHandler.MockStatusCode = HttpStatusCode.InternalServerError;
                 var syncResult = await _functionsSyncManager.TrySyncTriggersAsync(checkHash: true);
